changed price type to float64 to more accurately represent price

+++ internal/db/models/sku.go
@@ -3,7 +3,7 @@ package models
 type Sku struct {
 	SkuId string
 	Name string
-	Price int64
+	Price float64
 	Type string
 	Description string
 	ImageUrl string
